+++ src/scripts/bootstrap.js
@@ -2,6 +2,7 @@
  * Bootstrap Coffee Roulette Application
  */
 import "../scss/styles.scss"
+
 import {
   saveState,
   loadState,
@@ -23,7 +24,7 @@ import { initializeTooltips, setupGlobalTooltipHandler } from "./ui/tooltips.js"
 import { setupEventHandlers } from "./events/handlers.js"
 import { updateHistoryBadge } from "./utils/helpers.js"
 
-import { exportData, importCSVData } from "./data/exportCsv.js"
+import { exportData, importCSVData } from "./data/export-csv.js"
 
 if (typeof window !== "undefined") {
   loadState()
