+++ src/scripts/email/email-templates.js
