+++ src/scripts/pairing/pairing-helpers.js
